# [Random Reader Refactoring] Increase totalBytes for existing reader