### Description
- Inc totalBytes in case reading from existing reader as well.

### Link to the issue in case of a bug fix.
[b/426166667](https://b.corp.google.com/issues/426166667)

### Testing details
1. Manual - NA
2. Unit tests - Added
3. Integration tests - Done
4. Perf tests

Performance tests were conducted by averaging results from multiple runs.
  i. Scenario 1: 200 MiB file, 1 thread, 1 MiB block size
   - Without this fix: 180 MiB/s
   - With this fix: 250 MiB/s (same as master)

  ii. Scenario 2: 200 MiB file, 1 thread, 4 KiB block size
   - Without this fix: 20 MiB/s
   - With this fix: 50 MiB/s (same as master)

### Any backward incompatible change? If so, please explain.